move batch code into model specific helper modules.
